+++ test/builtins.js
@@ -1924,6 +1924,86 @@ async function testBuiltins() {
             }
         });
 
+        await ft.test([ft.list(ft.int(), 0, 10), ft.int(-10, 15)], `
+        testing list_drop
+        func main(a: []Int, n: Int) -> []Int {
+            a.drop(n)
+        }`, ([lst_, n_], res) => {
+            const lst = asIntList(lst_);
+            const n = Number(asInt(n_));
+            
+
+            if (n > lst.length) {
+                return isError(res, "empty list");
+            } else if (n < 0) {
+                return isError(res, "negative n in drop");
+            } else {
+                const resLst = asIntList(res);
+                
+                return ((n + resLst.length) == lst.length) && equalsList(resLst, lst.slice(n));
+            }
+        });
+
+        await ft.test([ft.list(ft.int(), 0, 10), ft.int(-10, 15)], `
+        testing list_take_end
+        func main(a: []Int, n: Int) -> []Int {
+            a.take_end(n)
+        }`, ([lst_, n_], res) => {
+            const lst = asIntList(lst_);
+            const n = Number(asInt(n_));
+            
+
+            if (n > lst.length) {
+                return isError(res, "list too short");
+            } else if (n < 0) {
+                return isError(res, "negative n in take_end");
+            } else {
+                const resLst = asIntList(res);
+                
+                return (n == resLst.length) && equalsList(resLst, lst.slice(lst.length - n));
+            }
+        });
+
+        await ft.test([ft.list(ft.int(), 0, 10), ft.int(-10, 15)], `
+        testing list_take
+        func main(a: []Int, n: Int) -> []Int {
+            a.take(n)
+        }`, ([lst_, n_], res) => {
+            const lst = asIntList(lst_);
+            const n = Number(asInt(n_));
+            
+
+            if (n > lst.length) {
+                return isError(res, "empty list");
+            } else if (n < 0) {
+                return isError(res, "negative n in take");
+            } else {
+                const resLst = asIntList(res);
+                
+                return (n == resLst.length) && equalsList(resLst, lst.slice(0, n));
+            }
+        });
+
+        await ft.test([ft.list(ft.int(), 0, 10), ft.int(-10, 15)], `
+        testing list_drop_end
+        func main(a: []Int, n: Int) -> []Int {
+            a.drop_end(n)
+        }`, ([lst_, n_], res) => {
+            const lst = asIntList(lst_);
+            const n = Number(asInt(n_));
+            
+
+            if (n > lst.length) {
+                return isError(res, "list too short");
+            } else if (n < 0) {
+                return isError(res, "negative n in drop_end");
+            } else {
+                const resLst = asIntList(res);
+                
+                return (n + resLst.length == lst.length) && equalsList(resLst, lst.slice(0, lst.length - n));
+            }
+        });
+
         await ft.test([ft.list(ft.int()), ft.int()], `
         testing list_prepend
         func main(a: []Int, b: Int) -> Bool {